PushpinDashboard/source/requirements.txt
aioredis==1.1.0
amqp==2.3.2
asgiref==2.3.2
async-timeout==3.0.0
attrs==18.1.0
autobahn==18.8.1
Automat==0.7.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
channels==2.1.3
channels-redis==2.3.0
chardet==3.0.4
constantly==15.1.0
daphne==2.2.2
Django==2.1
django-grip==1.8.0
djangorestframework==3.8.2
gripcontrol==3.2.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
kombu==4.2.1
msgpack==0.5.6
psycopg2-binary==2.7.5
pubcontrol==2.4.2
PyHamcrest==1.9.0
PyJWT==1.6.4
pytz==2018.5
pyzmq==17.1.2
requests==2.19.1
six==1.11.0
tnetstring3==0.3.1
Twisted==18.7.0
txaio==18.7.1
urllib3==1.23
vine==1.1.4
zope.interface==4.5.0
